refactor: new create changefeed cmd
