+++ pkg/cmd/cli/cli_changefeed_helper.go
@@ -1,0 +1,110 @@
+// Copyright 2021 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package cli
+
+import (
+	"fmt"
+	"strings"
+	"time"
+
+	"github.com/pingcap/errors"
+	"github.com/pingcap/ticdc/cdc/entry"
+	"github.com/pingcap/ticdc/cdc/kv"
+	"github.com/pingcap/ticdc/cdc/model"
+	"github.com/pingcap/ticdc/pkg/config"
+	"github.com/pingcap/ticdc/pkg/filter"
+	"github.com/pingcap/ticdc/pkg/security"
+	"github.com/spf13/cobra"
+	"github.com/tikv/client-go/v2/oracle"
+)
+
+const (
+	// tsGapWarning specifies the OOM threshold.
+	// 1 day in milliseconds
+	tsGapWarning = 86400 * 1000
+)
+
+// confirmLargeDataGap checks if a large data gap is used.
+func confirmLargeDataGap(cmd *cobra.Command, currentPhysical int64, startTs uint64) error {
+	tsGap := currentPhysical - oracle.ExtractPhysical(startTs)
+
+	if tsGap > tsGapWarning {
+		cmd.Printf("Replicate lag (%s) is larger than 1 days, "+
+			"large data may cause OOM, confirm to continue at your own risk [Y/N]\n",
+			time.Duration(tsGap)*time.Millisecond,
+		)
+		var yOrN string
+		_, err := fmt.Scan(&yOrN)
+		if err != nil {
+			return err
+		}
+		if strings.ToLower(strings.TrimSpace(yOrN)) != "y" {
+			return errors.NewNoStackError("abort changefeed create or resume")
+		}
+	}
+
+	return nil
+}
+
+// confirmIgnoreIneligibleTables confirm if user need to ignore ineligible tables.
+func confirmIgnoreIneligibleTables(cmd *cobra.Command) error {
+	cmd.Printf("Could you agree to ignore those tables, and continue to replicate [Y/N]\n")
+	var yOrN string
+	_, err := fmt.Scan(&yOrN)
+	if err != nil {
+		return err
+	}
+	if strings.ToLower(strings.TrimSpace(yOrN)) != "y" {
+		cmd.Printf("No changefeed is created because you don't want to ignore some tables.\n")
+		return errors.NewNoStackError("abort changefeed create or resume")
+	}
+
+	return nil
+}
+
+// getTables returns ineligibleTables and eligibleTables by filter.
+func getTables(cliPdAddr string, credential *security.Credential, cfg *config.ReplicaConfig, startTs uint64) (ineligibleTables, eligibleTables []model.TableName, err error) {
+	kvStore, err := kv.CreateTiStore(cliPdAddr, credential)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	meta, err := kv.GetSnapshotMeta(kvStore, startTs)
+	if err != nil {
+		return nil, nil, errors.Trace(err)
+	}
+
+	filter, err := filter.NewFilter(cfg)
+	if err != nil {
+		return nil, nil, errors.Trace(err)
+	}
+
+	snap, err := entry.NewSingleSchemaSnapshotFromMeta(meta, startTs, false /* explicitTables */)
+	if err != nil {
+		return nil, nil, errors.Trace(err)
+	}
+
+	for _, tableInfo := range snap.Tables() {
+		if filter.ShouldIgnoreTable(tableInfo.TableName.Schema, tableInfo.TableName.Table) {
+			continue
+		}
+		if !tableInfo.IsEligible(false /* forceReplicate */) {
+			ineligibleTables = append(ineligibleTables, tableInfo.TableName)
+		} else {
+			eligibleTables = append(eligibleTables, tableInfo.TableName)
+		}
+	}
+
+	return
+}
